+++ support-frontend/assets/pages/[countryGroupId]/checkout/helpers/getProductFields.ts
